services/eval-service/requirements.txt
anthropic==0.45.0
asyncpg==0.30.0
google-cloud-bigquery==3.27.0
pydantic==2.12.5
pydantic-settings==2.7.0
python-dotenv==1.2.2
ragas==0.1.21
sqlalchemy[asyncio]==2.0.48

# Test dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
